split out timeline tests
